requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
bitarray==2.9.2
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.2
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.32
cytoolz==0.12.3
eth-account==0.11.2
eth-bloom==3.0.1
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-tester==0.11.0b2
eth-typing==4.2.3
eth-utils==4.1.1
eth_abi==5.1.0
fonttools==4.53.0
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.7
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lru-dict==1.2.0
matplotlib==3.9.0
mazelib==0.9.14
multidict==6.0.5
numpy==1.23.4
packaging==23.2
parsimonious==0.10.0
pillow==10.3.0
protobuf==5.26.1
py-ecc==7.0.1
py-evm==0.10.1b1
py-solc-x==2.0.2
pycryptodome==3.20.0
pyparsing==3.1.2
PySMT==0.9.5
python-dateutil==2.9.0.post0
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rlp==4.0.1
rpds-py==0.18.1
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
toolz==0.12.1
trie==3.0.1
typing_extensions==4.11.0
urllib3==2.2.1
web3==6.19.0
websockets==12.0
yarl==1.9.4
